refactor: Moved HomeActivity to ui/house directory
